Fixed ConversationLogMiddleware and regestered the service with the bot.
