+++ src/Apprentice.Bot.Connectors/Middleware/CosmosFeedbackProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace ESFA.DAS.ProvideFeedback.Apprentice.Bot.Connectors.Middleware
+{
+    using ESFA.DAS.ProvideFeedback.Apprentice.Data;
+
+    public class CosmosFeedbackProvider
+    {
+        private readonly IDataRepository conversationLog;
+    }
+}
